Update project config
