+++ eslint.config.js
@@ -1,14 +1,12 @@
-import { includeIgnoreFile } from '@eslint/compat';
+import path from 'node:path';
 import js from '@eslint/js';
 import prettier from 'eslint-config-prettier';
 import svelte from 'eslint-plugin-svelte';
-import { defineConfig } from 'eslint/config';
+import { defineConfig, includeIgnoreFile } from 'eslint/config';
 import globals from 'globals';
-import { fileURLToPath } from 'node:url';
 import ts from 'typescript-eslint';
-import svelteConfig from './svelte.config.js';
 
-const gitignorePath = fileURLToPath(new URL('./.gitignore', import.meta.url));
+const gitignorePath = path.resolve(import.meta.dirname, '.gitignore');
 
 export default defineConfig(
 	includeIgnoreFile(gitignorePath),
@@ -23,7 +21,7 @@ export default defineConfig(
 		},
 		rules: {
 			// typescript-eslint strongly recommend that you do not use the no-undef lint rule on TypeScript projects.
-			// see: https://typescript-eslint.io/troubleshooting/faqs/eslint/#i-get-errors-from-the-no-undef-rule-about-global-variables-not-being-defined-even-though-there-are-no-typescript-errors
+			// see: https://typescript-eslint.io/troubleshooting/faqs/eslint/#i-get-errors-from-the-no-undef-rule-about-global-variables-not-being-defined-even-though-there-are-no-typescript-errors.
 			'no-undef': 'off',
 			// TODO: Keep an eye on https://github.com/sveltejs/eslint-plugin-svelte/issues/1353.
 			'svelte/no-navigation-without-resolve': [
@@ -40,14 +38,13 @@ export default defineConfig(
 			parserOptions: {
 				projectService: true,
 				extraFileExtensions: ['.svelte'],
-				parser: ts.parser,
-				svelteConfig
+				parser: ts.parser
 			}
 		}
 	},
 	{
 		// Override or add rule settings here, such as:
-		// 'svelte/button-has-type': 'error'
+		// 'svelte/button-has-type': 'error'.
 		rules: {}
 	}
 );
