+++ package.json
@@ -18,7 +18,6 @@
 	"devDependencies": {
 		"@content-collections/core": "^0.15.2",
 		"@content-collections/vite": "^0.3.0",
-		"@eslint/compat": "^2.1.0",
 		"@eslint/js": "^10.0.1",
 		"@fontsource/lexend": "^5.3.0",
 		"@fontsource/spectral": "^5.3.0",
@@ -45,7 +44,7 @@
 		"prettier": "^3.9.6",
 		"prettier-plugin-svelte": "^4.1.1",
 		"prettier-plugin-tailwindcss": "^0.8.1",
-		"shiki": "^4.4.1",
+		"shiki": "^4.4.2",
 		"simple-icons": "^16.28.0",
 		"svelte": "^5.56.8",
 		"svelte-check": "^4.7.4",
@@ -53,7 +52,7 @@
 		"tailwind-merge": "^3.6.0",
 		"tailwindcss": "^4.3.3",
 		"typescript": "^6.0.3",
-		"typescript-eslint": "^8.65.0",
+		"typescript-eslint": "^8.66.0",
 		"vite": "^8.2.0",
 		"zod": "^4.4.3"
 	}
