+++ prettier.config.js
@@ -1,0 +1,20 @@
+/** @type {import('prettier').Config} */
+const config = {
+	useTabs: true,
+	singleQuote: true,
+	trailingComma: 'none',
+	printWidth: 100,
+	proseWrap: 'always',
+	plugins: ['prettier-plugin-svelte', 'prettier-plugin-tailwindcss'],
+	tailwindStylesheet: './src/app.css',
+	overrides: [
+		{
+			files: '*.svelte',
+			options: {
+				parser: 'svelte'
+			}
+		}
+	]
+};
+
+export default config;
